+++ src/layers/submarineCables/policy.js
@@ -78,7 +78,15 @@ export const CABLE_LABEL_DEPTH_DECISION = Object.freeze({
  */
 
 export const CABLE_GLOBE_STACK_IDS = Object.freeze(
-  new Set(['bing-aerial', 'bing-labels', 'esri-imagery', 'osm']),
+  new Set([
+    'bing-aerial',
+    'bing-labels',
+    'esri-imagery',
+    'osm',
+    'carto-positron',
+    'carto-dark',
+    'openfreemap-dark',
+  ]),
 );
 
 /** EntityCluster's private marker collections and their required types. */