+++ src/mapStackChips.js
@@ -19,6 +19,9 @@ export const PRESENTED_MAP_STACK_IDS = Object.freeze([
   'bing-labels',
   'esri-imagery',
   'osm',
+  'carto-positron',
+  'carto-dark',
+  'openfreemap-dark',
 ]);
 
 /**